requirements/common.txt
Mezzanine==4.0.1
gunicorn
psycopg2
django-compressor
